# backend/openapi-schema.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/games/{id}": {
      "put": {
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGameDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The game has been successfully updated."
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CRUDErrorBadRequestResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update game name",
        "tags": [
          "games"
        ]
      },
      "delete": {
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The game has been successfully deleted."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a game",
        "tags": [
          "games"
        ]
      }
    },
...
    "/api/games/{idOrShortUrl}": {
      "get": {
        "operationId": "findOne",
        "parameters": [
          {
            "name": "idOrShortUrl",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get game data by ID or short URL",
        "tags": [
          "games"
        ]
      }
    },
    "/api/games/{gameId}/scenes": {
      "post": {
        "operationId": "create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGameSceneDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The scene has been successfully created."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new game scene",
        "tags": [
          "gamesScenes"
        ]
      },
      "get": {
        "operationId": "findAll",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get paginated list of game scenes",
        "tags": [
          "gamesScenes"
        ]
      }
    },
    "/api/games/{gameId}/scenes/active": {
      "get": {
        "operationId": "findActive",
        "parameters": [
          {
            "name": "gameId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The active scene.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameSceneEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get active scene by game id",
        "tags": [
          "gamesScenes"
        ]
      }
    },
    "/api/games/{gameId}/scenes/{id}": {
      "get": {
        "operationId": "findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get full scene by id",
        "tags": [
          "gamesScenes"
        ]
      },
      "patch": {
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGameSceneDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update scene metadata and grid settings",
        "tags": [
          "gamesScenes"
        ]
      },
      "delete": {
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a scene",
        "tags": [
          "gamesScenes"
        ]
      }
    },
    "/api/games/{gameId}/scenes/{sceneId}/layers": {
      "post": {
        "operationId": "createLayer",
        "parameters": [
          {
            "name": "sceneId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSceneLayerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new layer in a scene",
        "tags": [
          "gamesScenes"
        ]
      },
      "get": {
        "operationId": "findLayers",
        "parameters": [
          {
            "name": "sceneId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List layers for a scene",
        "tags": [
          "gamesScenes"
        ]
      }
    },
    "/api/games/{gameId}/scenes/{sceneId}/layers/{layerId}": {
      "patch": {
        "operationId": "updateLayer",
        "parameters": [
          {
            "name": "sceneId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "layerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSceneLayerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update layer",
        "tags": [
          "gamesScenes"
        ]
      },
      "delete": {
        "operationId": "deleteLayer",
        "parameters": [
          {
            "name": "sceneId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "layerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete layer",
        "tags": [
          "gamesScenes"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Documentation",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "FieldError": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "description": "Form field name"
          },
          "message": {
            "type": "string",
            "description": "Form field error message"
          }
        },
        "required": [
          "field",
          "message"
        ]
      },
      "CRUDErrorBadRequestResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "description": "error status code"
          },
          "message": {
            "type": "string",
            "description": "error message"
          },
          "errors": {
            "description": "Error messages by fields",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          },
          "timestamp": {
            "format": "date-time",
            "type": "string",
            "description": "timestamp"
          }
        },
        "required": [
          "statusCode",
          "message",
          "errors",
          "timestamp"
        ]
      },
      "UpdateGameDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "New Game Name"
          }
        },
        "required": [
          "name"
        ]
      },
      "GameResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "shortUrl": {
            "type": "string",
            "example": "abcdefgh"
          },
          "name": {
            "type": "string",
            "example": "My Awesome Game"
          },
          "status": {
            "type": "string",
            "example": "CREATED",
            "enum": [
              "CREATED",
              "STARTED"
            ]
          }
        },
        "required": [
          "id",
          "shortUrl",
          "name",
          "status"
        ]
      },
      "CreateGameSceneDto": {
        "type": "object",
        "properties": {
          "gameId": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "stageWidth": {
            "type": "number"
          },
          "stageHeight": {
            "type": "number"
          },
          "backgroundColor": {
            "type": "string"
          },
          "gridType": {
            "type": "string",
            "enum": [
              "square"
            ]
          },
          "gridCellSize": {
            "type": "number"
          },
          "gridMetricSystem": {
            "type": "string",
            "enum": [
              "Feet",
              "Meters",
              "Miles",
              "Squares"
            ]
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "required": [
          "gameId",
          "name"
        ]
      },
      "GameSceneLayerEntity": {
        "type": "object",
        "properties": {}
      },
      "GameSceneEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "name": {
            "type": "string",
            "example": "Living Room"
          },
          "stageJSON": {
            "type": "object"
          },
          "stageWidth": {
            "type": "number",
            "example": 800
          },
          "stageHeight": {
            "type": "number",
            "example": 600
          },
          "backgroundColor": {
            "type": "string",
            "example": "#ffffff"
          },
          "gridType": {
            "type": "string",
            "enum": [
              "square"
            ],
            "default": "square"
          },
          "gridCellSize": {
            "type": "number",
            "example": 70,
            "default": 70
          },
          "gridMetricSystem": {
            "type": "string",
            "enum": [
              "Feet",
              "Meters",
              "Miles",
              "Squares"
            ],
            "default": "Squares"
          },
          "isActive": {
            "type": "boolean",
            "default": true
          },
          "layers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GameSceneLayerEntity"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "gameId": {
            "type": "number",
            "example": 1
          }
        },
        "required": [
          "id",
          "name",
          "stageJSON",
          "stageWidth",
          "stageHeight",
          "backgroundColor",
          "gridType",
          "gridCellSize",
          "gridMetricSystem",
          "isActive",
          "layers",
          "createdAt",
          "updatedAt",
          "gameId"
        ]
      },
      "UpdateGameSceneDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "stageWidth": {
            "type": "number"
          },
          "stageHeight": {
            "type": "number"
          },
          "backgroundColor": {
            "type": "string"
          },
          "gridType": {
            "type": "string",
            "enum": [
              "square"
            ]
          },
          "gridCellSize": {
            "type": "number"
          },
          "gridMetricSystem": {
            "type": "string",
            "enum": [
              "Feet",
              "Meters",
              "Miles",
              "Squares"
            ]
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
      "CreateSceneLayerDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "isLocked": {
            "type": "boolean"
          },
          "isVisible": {
            "type": "boolean"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateSceneLayerDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "isLocked": {
            "type": "boolean"
          },
          "isVisible": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}