# backend/src/modules/gameScene/presentation/rest/gameScene.controller.ts
import { Controller, Post, Body, Get, Param, Patch, Delete, Query, UseGuards, HttpStatus } from '@nestjs/common';
import { CommandBus, QueryBus } from '@nestjs/cqrs';
import { ApiTags, ApiOperation, ApiResponse, ApiBearerAuth } from '@nestjs/swagger';
import { CreateGameSceneDto, UpdateGameSceneDto, CreateSceneLayerDto, UpdateSceneLayerDto } from './dtos/gameScene.dtos';
import { CreateGameSceneCommand } from '../../application/commands/impl/createGameScene.command';
import { UpdateGameSceneCommand } from '../../application/commands/impl/updateGameScene.command';
import { DeleteGameSceneCommand } from '../../application/commands/impl/deleteGameScene.command';
import { CreateSceneLayerCommand } from '../../application/commands/impl/createSceneLayer.command';
import { UpdateSceneLayerCommand } from '../../application/commands/impl/updateSceneLayer.command';
import { DeleteSceneLayerCommand } from '../../application/commands/impl/deleteSceneLayer.command';
import {
  GetGameScenesQuery,
  GetGameSceneByIdQuery,
  GetSceneLayersQuery,
  GetActiveGameSceneByGameIdQuery,
} from '../../application/queries/impl/gameScene.queries';
import { JwtAuthGuard } from '../../../account/auth/guards/jwtAuth.guard';
import { GameSceneEntity } from '../../infrastructure/persistence/typeorm/entities/gameScene.entity';

@ApiTags('gamesScenes')
@ApiBearerAuth()
@UseGuards(JwtAuthGuard)
@Controller('games/:gameId/scenes')
export class GameSceneController {
  constructor(
    private readonly commandBus: CommandBus,
    private readonly queryBus: QueryBus,
  ) {}

  @Post()
  @ApiOperation({ summary: 'Create a new game scene' })
  @ApiResponse({ status: 201, description: 'The scene has been successfully created.' })
  async create(@Body() dto: CreateGameSceneDto) {
    return this.commandBus.execute(
      new CreateGameSceneCommand(
        dto.gameId,
        dto.name,
        dto.stageWidth,
        dto.stageHeight,
        dto.backgroundColor,
        dto.gridType,
        dto.gridCellSize,
        dto.gridMetricSystem,
        dto.isActive,
      ),
    );
  }

  @Get()
  @ApiOperation({ summary: 'Get paginated list of game scenes' })
  async findAll(@Query('page') page: number = 1, @Query('limit') limit: number = 10) {
    return this.queryBus.execute(new GetGameScenesQuery(Number(page), Number(limit)));
  }

  @Get('active')
  @ApiOperation({ summary: 'Get active scene by game id' })
  @ApiResponse({ status: HttpStatus.OK, type: GameSceneEntity, description: 'The active scene.' })
  async findActive(@Param('gameId') gameId: string) {
    return this.queryBus.execute(new GetActiveGameSceneByGameIdQuery(Number(gameId)));
  }

  @Get(':id')
  @ApiOperation({ summary: 'Get full scene by id' })
  async findOne(@Param('id') id: string) {
    return this.queryBus.execute(new GetGameSceneByIdQuery(id));
  }

  @Patch(':id')
  @ApiOperation({ summary: 'Update scene metadata and grid settings' })
  async update(@Param('id') id: string, @Body() dto: UpdateGameSceneDto) {
    return this.commandBus.execute(
      new UpdateGameSceneCommand(
        id,
        dto.name,
        dto.stageWidth,
        dto.stageHeight,
        dto.backgroundColor,
        dto.gridType,
        dto.gridCellSize,
        dto.gridMetricSystem,
        dto.isActive,
      ),
    );
  }

  @Delete(':id')
  @ApiOperation({ summary: 'Delete a scene' })
  async delete(@Param('id') id: string) {
    return this.commandBus.execute(new DeleteGameSceneCommand(id));
  }

  @Post(':sceneId/layers')
  @ApiOperation({ summary: 'Create a new layer in a scene' })
  async createLayer(@Param('sceneId') sceneId: string, @Body() dto: CreateSceneLayerDto) {
    return this.commandBus.execute(new CreateSceneLayerCommand(sceneId, dto.name, dto.isLocked, dto.isVisible));
  }

  @Get(':sceneId/layers')
  @ApiOperation({ summary: 'List layers for a scene' })
  async findLayers(@Param('sceneId') sceneId: string) {
    return this.queryBus.execute(new GetSceneLayersQuery(sceneId));
  }

  @Patch(':sceneId/layers/:layerId')
  @ApiOperation({ summary: 'Update layer' })
  async updateLayer(@Param('sceneId') sceneId: string, @Param('layerId') layerId: string, @Body() dto: UpdateSceneLayerDto) {
    return this.commandBus.execute(new UpdateSceneLayerCommand(sceneId, layerId, dto.name, dto.isLocked, dto.isVisible));
  }

  @Delete(':sceneId/layers/:layerId')
  @ApiOperation({ summary: 'Delete layer' })
  async deleteLayer(@Param('sceneId') sceneId: string, @Param('layerId') layerId: string) {
    return this.commandBus.execute(new DeleteSceneLayerCommand(sceneId, layerId));
  }
}